backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
email-validator==2.3.0
